# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/notifications": {
      "get": {
        "operationId": "listNotifications",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationsResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/v1/notifications/{id}/read": {
      "patch": {
        "operationId": "markNotificationRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkReadResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "415": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "",
        "tags": [
          "notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "Meeting Room Booking API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mrb_session"
      }
    },
    "schemas": {
...
      "ApiErrorDetailsDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "example": "Request validation failed"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          },
          "requestId": {
            "type": "string",
            "example": "b6a3a1f7-07d8-4fd8-a0c5-ff1f5aa9b568"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ]
      },
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetailsDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "NotificationItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "type": {
            "type": "string",
            "example": "NEXT_BOOKING_STARTS"
          },
          "message": {
            "type": "string",
            "example": "«Стендап» у Берлін завершується за 10 хв — наступне бронювання починається одразу"
          },
          "roomName": {
            "type": "string",
            "example": "Берлін"
          },
          "currentBookingId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174001"
          },
          "nextBookingId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174002"
          },
          "scheduledFor": {
            "type": "string",
            "example": "2026-08-03T10:50:00.000Z"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-08-03T10:50:01.000Z"
          },
          "readAt": {
            "type": "object",
            "example": "2026-08-03T10:55:00.000Z",
            "nullable": true
          }
        },
        "required": [
          "id",
          "type",
          "message",
          "roomName",
          "currentBookingId",
          "nextBookingId",
          "scheduledFor",
          "createdAt",
          "readAt"
        ]
      },
      "NotificationsResponseDto": {
        "type": "object",
        "properties": {
          "notifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationItemDto"
            }
          },
          "unreadCount": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "notifications",
          "unreadCount"
        ]
      },
      "NotificationReadSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "readAt": {
            "type": "string",
            "example": "2026-08-03T10:55:00.000Z"
          }
        },
        "required": [
          "id",
          "readAt"
        ]
      },
      "MarkReadResponseDto": {
        "type": "object",
        "properties": {
          "notification": {
            "$ref": "#/components/schemas/NotificationReadSummaryDto"
          }
        },
        "required": [
          "notification"
        ]
      }
    }
  }
}